moved format tags in one location
